Recover bool and byte predicate boundaries
Classify the first bool/char/BOOL tranche from retail listings, fix the affected virtual signatures and byte-storage contracts, and record the evidence in a dedicated boundary inventory.\n\nRecover void DrawImmediate dispatch, exact bool toggle predicates, byte refresh arguments, and explicit logical-to-integer conversions.\n\nVerified with just agent-check and just precommit. Stats: 7 functions improved, 3 newly exact, no regressions.

+++ src/game/TArmyPlayer.cpp
@@ -65,7 +65,8 @@ short __cdecl CompareTacticalCursorEntriesByActionClassPriority(void* a, void* b
   if (priorityA < priorityB) {
     return 1;
   }
-  return -(priorityA != priorityB); // neg/sbb idiom: -1 when priorityA > priorityB, else 0
+  // Explicitly cross from the logical comparison into the integer comparator domain.
+  return static_cast<short>(-static_cast<int>(priorityA != priorityB));
 }
 // SYNTHETIC: IMPERIALISM 0x0059b110
 // TArmyPlayer::CreateObject
